// src/app/components/interfaces/schedule-inter-data.ts
export interface ScheduleInterData {
    num: number;
    id: string;
    date: string;
    time: string;
    awayTeamCity: string;
    awayTeamName: string;
    awayTeamAbbreviation: string;
    homeTeamCity: string;
    homeTeamName: string;
    homeTeamAbbreviation: string;
    location: string;
}